pyproject.toml
[project]
name = "django-skeleton"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Duy Nguyen", email = "duynguyenngoc@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "django>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "python-decouple>=3.8",
    "whitenoise>=6.6.0",
    "gunicorn>=21.2.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "pre-commit>=4.0.1",
]

# --- START Ruff Configuration ---
[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4
# Enable fix behavior by-default when running ruff
fix = true
# Enumerate all fixed violations.
show-fixes = true
# Exclude directories (e.g., virtual environments, build artifacts)
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "**/migrations/**",
]
[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors (PEP 8)
  "F",    # pyflakes (general errors)
  "I",    # isort (import sorting)
  "D",    # pydocstyle (docstring conventions)
  "UP",   # pyupgrade (modern syntax)
  "ANN",  # flake8-annotations (type hints)
  "B",    # flake8-bugbear (best practices)
  "N",    # pep8-naming
  "C90",  # mccabe (complexity)
]
# Ignore specific rules if needed (example: ignore missing docstrings for __init__)
ignore = [
  "D100",  # Missing docstring in public module
  "D104",  # Missing docstring in public package
  # Add more as appropriate for your codebase
]
# Formatting options
[tool.ruff.format]
# Like Black, use double quotes for strings
quote-style = "double"
# Like Black, indent with spaces, rather than tabs
indent-style = "space"
# Like Black, respect magic trailing commas
skip-magic-trailing-comma = false
# Use `\n` line endings for all files
line-ending = "lf"
# Enable reformatting of code snippets in docstrings
docstring-code-format = true
# Docstring style: Google
[tool.ruff.lint.pydocstyle]
convention = "google"
[tool.ruff.lint.flake8-annotations]
# Suppress ANN401 for dynamically typed *args and **kwargs arguments
allow-star-arg-any = true
# Allow the omission of a return type hint for __init__ if at least one argument is annotated
mypy-init-return = true
# Suppress ANN000-level violations for arguments matching the "dummy" variable regex (like _)
suppress-dummy-args = true
# Suppress ANN200-level violations for functions
suppress-none-returning = true
[tool.ruff.lint.isort]
# Whether to place import from imports before straight imports when sorting
from-first = false
# --- END Ruff Configuration ---